requirements.txt
# FastAPI + server
fastapi==0.111.0         
uvicorn[standard]==0.30.1 

# LangChain ecosystem
langchain==0.2.6                 
langchain-community==0.2.6       
langchain-openai==0.1.13         
openai==1.35.3                   

# Embeddings (CLIP)
transformers==4.42.3              
torch==2.3.1                      
Pillow==10.3.0                   

# Vector store
faiss-cpu==1.8.0                  

# Caching
redis==5.0.7                      

# Database
psycopg2-binary==2.9.9            
asyncpg==0.29.0                  
sqlalchemy==2.0.31               

# Utilities
numpy==1.26.4                  
python-dotenv==1.0.1             
pydantic==2.7.4                   
httpx==0.27.0                   

# Testing
pytest==8.2.2                
pytest-asyncio==0.23.7           